Prevent error if get_current_screen() not defined
Per the WordPress Codex: get_current_screen() is defined on most admin pages, but not all. Thus there are cases where is_admin() will return true, but attempting to call get_current_screen() will result in a fatal error because it is not defined... Checking that the function exists before calling it allows us to avoid the error, but we can no longer be sure that we are editing a post or page and can only assume that we are not.